Add files via upload
Excited to share that I’ve developed a resume generator 📄as part of my web developer internship at Coding Raja Technologies!  
This project involved creating a dynamic and user-friendly tool that helps users easily build professional resumes. It was a rewarding experience that enhanced my skills in both front-end and back-end development. 
Grateful for the opportunity to work on impactful solutions, and I’m eager to tackle more challenges in the future!"